move pt_PT before pt_BR in language list
.po files are transformed to .tr files, and then gfxboot uses the name
of the .tr file as a key to lookup in a language list to find a pretty
name for the language. However, most of the languages in the language
list are of the form xx_YY while the .tr file are mostly of the xx.tr
form, so gfxboot code truncates the langage name when doing the lookup.
For portuguese, we've got both pt.tr and pt_BR.tr and the language list
had both pt_PT and pt_BR. However, pt_BR is listed before pt_PT, so when
looking up the pretty name for pt.tr, the pt_BR entry is used, leading
to duplicate "pt_BR" in the language menu instead of having both pt and
pt_BR. Moving pt_PT up in the list should workaround this issue

git-svn-id: svn+ssh://svn.mandriva.com/svn/soft/theme/mandriva-gfxboot-theme/trunk@269518 99302b65-d5f7-0310-b3dd-f8cd6f4e3d94